@@ -49,7 +49,7 @@ setup(
     packages=find_packages('src'),
     package_dir={'': 'src'},
     include_package_data=True,
-    python_requires=">=3.7",
+    python_requires=">=3.8",
     install_requires=[
         'setuptools',
         'funcparserlib>=1.0.0a0',
